Registration
The finalized registration system, without the mailtrap section
