+++ routes/web.php
@@ -5,8 +5,6 @@ use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\PostController;
 use App\Actions\Fortify\CreateNewUser;
 
-use App\Http\Controllers\RegisterStepTwoController;
-use App\Http\Controllers\RegisterStepOneController;
 
 
 /*
@@ -26,28 +24,6 @@ Route::get('/blog', [PostController::class, 'index'])->name('posts.index');
 
 Route::get('/blog/{post:slug}', [PostController::class, 'show'])->name('posts.show');
 
-//Route::post('/registerUser', [RegisterStepOneController::class, 'toResponse']); 
-//Route::get('/registerUser', [RegisterStepOneController::class, 'toResponse']); 
 
-Route::post('/register', [RegisterStepOneController::class, 'toResponse'])->name('register');
-
-Route::post('/register2', [RegisterStepTwoController::class, 'store'])->name('register2.post');
-Route::get('/register2', [RegisterStepTwoController::class, 'create'])->name('register2.create');
-
-//Route::group(['middleware'=> [
-//    'auth',
-//    'verified',
-//  ]], function () {
-//  
-//    Route::group(['middleware'=>['registration_completed']], function() {
-//      
-//      
-//    });
-//  
-//    // Adding route to second layer of registration
-//    //Route::get('/register2', [RegisterStepTwoController::class, 'create'])->name('register2.create');
-//    //Route::post('/register2', [RegisterStepTwoController::class, 'store'])->name('register2.post');
-//  });
-//
